pyproject.toml
[tool.poetry]
name = "cvm_agents"
version = "0.1.0"
description = "CVM (Customer Value Management) System for Telecommunications"
authors = ["Tarek Salhi <tareksalhi0@gmail.com>"]
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pandas = "^2.2.3"
numpy = "^1.21.2"
tabulate = "^0.8.9"
colorama = "^0.4.4"
langchain-core = ">=0.2.43,<0.3.0"
typer = "^0.4.0"
python-dateutil = "^2.8.2"
langgraph = "^0.2.67"
openai = "^1.60.0"
python-dotenv = "^1.0.1"
smolagents = "^1.7.0"
litellm = "^1.60.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"